collab work with Brent and Caity for Sunday night, implements all mongo routes and does some readme, and modularizes route files
