+++ commands.txt
@@ -4,6 +4,10 @@ testing commands:
 GET
 http :8080/api/book
 
+GET ONE
+5c394f339bdd10cb6501c38c
+
+
 POST
 echo '{  "title": "Dune","author":"Frank Herbert","isbn":"ISBN_13 9780441013593", "image_url":"http://books.google.com/books/content?id=B1hSG45JCX4C&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api", "description":"Follows the adventures of Paul Atreides, the son of a betrayed duke given up for dead on a treacherous desert planet and adopted by its fierce, nomadic people, who help him unravel his most unexpected destiny.", "bookshelf":"Fantasy"}' | http post :8080/api/book
