+++ src/models/book-constructor.js
@@ -1,0 +1,14 @@
+'use strict';
+
+function Book(info) {
+  let placeholderImage = 'https://i.imgur.com/J5LVHEL.jpg';
+
+  this.title = info.title || 'No title available';
+  // TODO: handle when the authors array doesn't exist
+  this.author = info.authors[0]|| 'No authors available'
+  this.isbn = info.industryIdentifiers ? `ISBN_13 ${info.industryIdentifiers[0].identifier}` : 'No ISBN available';
+  this.image_url = info.imageLinks ? info.imageLinks.smallThumbnail : placeholderImage;
+  this.description = info.description || 'No description available';
+};
+
+module.exports = Book;
