+++ src/routes/handlers/fetch-books-from-api.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = (req, res, next) => {
+  let url = 'https://www.googleapis.com/books/v1/volumes?q=';
+
+  if (req.body.search[1] === 'title') { url += `+intitle:${req.body.search[0]}`; }
+  if (req.body.search[1] === 'author') { url += `+inauthor:${req.body.search[0]}`; }
+
+  superagent.get(url)
+    .then(apiResponse => apiResponse.body.items.map(bookResult => new Book(bookResult.volumeInfo)))
+    .then(results => res.render('pages/searches/show', {results: results}))
+    .catch( next );
+}
