+++ src/routes/handlers/get-all-books.js
@@ -1,0 +1,17 @@
+'use strict';
+
+module.exports = (req, res, next) => {
+  books.get()
+    .then( data => {
+      const output = {
+        count: data.length,
+        results: data,
+      };
+      if (output.count){
+        res.render('pages/index', {books: output.results});
+      }else{
+        res.render('pages/searches/new');
+      }
+    })
+    .catch( next );
+}
