+++ src/routes/handlers/get-book.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = (req, res, next) => {
+  books.get(req.params.id)
+    .then( data => {
+      const output = {
+        count: data.length,
+        results: data,
+      };
+      res.render(`pages/books/show`, {book: output.results[0], bookshelves: []});
+    })
+    .catch( next );
+};
