+++ src/routes/handlers/new-search.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = (req, res) => {
+  res.render('pages/searches/new');
+}
