+++ src/routes/handlers/update-book.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = (req, res, next) => {
+  books.put(req.params.id, req.body)
+  .then( data => {
+    const output = {
+      count: data.length,
+      results: data,
+    };
+    res.redirect(`/books/${req.params.id}`);
+  })
+  .catch( next );
+}
