--- test.js
@@ -1,262 +1,0 @@
-{ info:
-  { title: 'Precise Leveling in New York City',
-    authors:
-     [ 'New York (N.Y.). Board of Estimate and Apportionment',
-       'Frederick W. Koop' ],
-    publishedDate: '1914',
-    industryIdentifiers: [ { type: 'OTHER', identifier: 'CHI:086729098' } ],
-    readingModes: { text: false, image: true },
-    pageCount: 263,
-    printType: 'BOOK',
-    categories: [ 'Bench-marks' ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: 'full-1.0.0',
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=MvAyAQAAMAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=MvAyAQAAMAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=MvAyAQAAMAAJ&printsec=frontcover&dq=inauthor:new+york&hl=&cd=1&source=gbs_api',
-    infoLink:
-     'https://play.google.com/store/books/details?id=MvAyAQAAMAAJ&source=gbs_api',
-    canonicalVolumeLink: 'https://market.android.com/details?id=book-MvAyAQAAMAAJ' } }
-{ info:
-  { title: 'Guide to New York City Landmarks',
-    authors:
-     [ 'Andrew Dolkart',
-       'Matthew A. Postal',
-       'New York City Landmarks Preservation Commission' ],
-    publisher: 'John Wiley & Sons',
-    publishedDate: '2004',
-    description:
-     'Five boroughs rich with history–and one guide to it all The official and only complete guide to New York City′s landmarks, this beautiful Third Edition has been updated to include 128 new individual sites and sixteen new historic districts. Mayor Michael Bloomberg opens this treasure trove of historical discovery with an elegant Foreword. What follows are insightful descriptions of more than 1,000 individual landmarks and 84 historic districts. Everyone will feel like a native New Yorker when they quickly point out landmarks with the help of 80 easy–to–read maps. This new edition also features new photographs, enhanced maps, and more than a dozen themed sections that make it easy to create a customized sightseeing experience. New York City Landmarks Preservation Commission (New York, NY) is the agency responsible for identifying and designating local landmarks and historic districts. Established in 1965 in response to the destruction of the original Penn Station, the agency is comprised of eleven members appointed by the mayor and a professional staff.',
-    industryIdentifiers:
-     [ { type: 'ISBN_10', identifier: '0471369004' },
-       { type: 'ISBN_13', identifier: '9780471369004' } ],
-    readingModes: { text: false, image: true },
-    pageCount: 406,
-    printType: 'BOOK',
-    categories: [ 'Architecture' ],
-    averageRating: 4,
-    ratingsCount: 1,
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: '0.0.1.0.preview.1',
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=W4wyfi6GG6gC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=W4wyfi6GG6gC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=W4wyfi6GG6gC&printsec=frontcover&dq=inauthor:new+york&hl=&cd=2&source=gbs_api',
-    infoLink:
-     'http://books.google.com/books?id=W4wyfi6GG6gC&dq=inauthor:new+york&hl=&source=gbs_api',
-    canonicalVolumeLink:
-     'https://books.google.com/books/about/Guide_to_New_York_City_Landmarks.html?hl=&id=W4wyfi6GG6gC' } }
-{ info:
-  { title: 'Documents of the Senate of the State of New York',
-    authors: [ 'New York (State). Legislature. Senate' ],
-    publishedDate: '1916',
-    industryIdentifiers: [ { type: 'OTHER', identifier: 'CORNELL:31924093482986' } ],
-    readingModes: { text: false, image: true },
-    printType: 'BOOK',
-    categories: [ 'Government publications' ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: '0.1.0.0.full.1',
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=XEA8AQAAMAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=XEA8AQAAMAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=XEA8AQAAMAAJ&printsec=frontcover&dq=inauthor:new+york&hl=&cd=3&source=gbs_api',
-    infoLink:
-     'https://play.google.com/store/books/details?id=XEA8AQAAMAAJ&source=gbs_api',
-    canonicalVolumeLink: 'https://market.android.com/details?id=book-XEA8AQAAMAAJ' } }
-{ info:
-  { title:
-     'Report of the Civil Service Commission of the State of New York',
-    authors: [ 'New York (State). Civil Service Commission' ],
-    publishedDate: '1903',
-    industryIdentifiers: [ { type: 'OTHER', identifier: 'HARVARD:LI2VFR' } ],
-    readingModes: { text: false, image: true },
-    printType: 'BOOK',
-    categories: [ 'Civil service' ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: 'full-1.0.0',
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=BUQVAAAAYAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=BUQVAAAAYAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=BUQVAAAAYAAJ&pg=PP1&dq=inauthor:new+york&hl=&cd=4&source=gbs_api',
-    infoLink:
-     'https://play.google.com/store/books/details?id=BUQVAAAAYAAJ&source=gbs_api',
-    canonicalVolumeLink: 'https://market.android.com/details?id=book-BUQVAAAAYAAJ' } }
-{ info:
-  { title: 'A Community of Many Worlds',
-    subtitle: 'Arab Americans in New York City',
-    authors: [ 'Ny Museum', 'Museum of the City of New York' ],
-    publisher: 'Syracuse University Press',
-    publishedDate: '2002',
-    description:
-     "New York City's main Arab communities exemplify the continuity and change that has taken place throughout the city's rich history. The Museum of the City of New York, in partnershipwith the Middle East Institute at Columbia University and a group of local Arab and non-Arab scholars, activists and educators, undertook a long overdue exploration of New York's Arab populations. The result is a revealing collection of writings and photographs that document and tell the stories of these communities.",
-    industryIdentifiers:
-     [ { type: 'ISBN_10', identifier: '0815607393' },
-       { type: 'ISBN_13', identifier: '9780815607397' } ],
-    readingModes: { text: false, image: true },
-    pageCount: 280,
-    printType: 'BOOK',
-    categories: [ 'History' ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: '0.1.2.0.preview.1',
-    panelizationSummary: { containsEpubBubbles: false, containsImageBubbles: false },
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=qFSIf5Zz26gC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=qFSIf5Zz26gC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=qFSIf5Zz26gC&printsec=frontcover&dq=inauthor:new+york&hl=&cd=5&source=gbs_api',
-    infoLink:
-     'http://books.google.com/books?id=qFSIf5Zz26gC&dq=inauthor:new+york&hl=&source=gbs_api',
-    canonicalVolumeLink:
-     'https://books.google.com/books/about/A_Community_of_Many_Worlds.html?hl=&id=qFSIf5Zz26gC' } }
-{ info:
-  { title:
-     'Index of Awards on Claims of the Soldiers of the War of 1812',
-    authors:
-     [ "Adjutant General's Office New York",
-       "New York (State) Adjutant General's Office" ],
-    publisher: 'Genealogical Publishing Com',
-    publishedDate: '2013-08',
-    industryIdentifiers:
-     [ { type: 'ISBN_13', identifier: '9780806302584' },
-       { type: 'ISBN_10', identifier: '0806302585' } ],
-    readingModes: { text: false, image: true },
-    pageCount: 584,
-    printType: 'BOOK',
-    categories: [ 'Reference' ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: '1.1.0.0.preview.1',
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=iEWOp6fYyhgC&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=iEWOp6fYyhgC&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=iEWOp6fYyhgC&pg=PA575&dq=inauthor:new+york&hl=&cd=6&source=gbs_api',
-    infoLink:
-     'https://play.google.com/store/books/details?id=iEWOp6fYyhgC&source=gbs_api',
-    canonicalVolumeLink: 'https://market.android.com/details?id=book-iEWOp6fYyhgC' } }
-{ info:
-  { title: 'Annual Exhibition - Architectural League of New York',
-    authors: [ 'Architectural League of New York' ],
-    publishedDate: '1898',
-    industryIdentifiers: [ { type: 'OTHER', identifier: 'UOM:39015069419870' } ],
-    readingModes: { text: false, image: true },
-    printType: 'BOOK',
-    categories: [ 'Architecture' ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: 'full-1.0.0',
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=6ww6AAAAMAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=6ww6AAAAMAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=6ww6AAAAMAAJ&printsec=frontcover&dq=inauthor:new+york&hl=&cd=7&source=gbs_api',
-    infoLink:
-     'https://play.google.com/store/books/details?id=6ww6AAAAMAAJ&source=gbs_api',
-    canonicalVolumeLink: 'https://market.android.com/details?id=book-6ww6AAAAMAAJ' } }
-{ info:
-  { title: 'New York City Charter and Administrative Code annotated',
-    subtitle:
-     'a complete text of the New York City charter and the New York City administrative code with court decisions from the time of the enactment of the code and charter',
-    authors: [ 'New York (N.Y.)', 'New York (N.Y.).' ],
-    publishedDate: '1989',
-    industryIdentifiers: [ { type: 'OTHER', identifier: 'STANFORD:36105062992040' } ],
-    readingModes: { text: false, image: false },
-    printType: 'BOOK',
-    categories: [ 'Law' ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: 'preview-1.0.0',
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=xClPAQAAIAAJ&printsec=frontcover&img=1&zoom=5&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=xClPAQAAIAAJ&printsec=frontcover&img=1&zoom=1&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=xClPAQAAIAAJ&q=inauthor:new+york&dq=inauthor:new+york&hl=&cd=8&source=gbs_api',
-    infoLink:
-     'http://books.google.com/books?id=xClPAQAAIAAJ&dq=inauthor:new+york&hl=&source=gbs_api',
-    canonicalVolumeLink:
-     'https://books.google.com/books/about/New_York_City_Charter_and_Administrative.html?hl=&id=xClPAQAAIAAJ' } }
-{ info:
-  { title:
-     'Catalogue of the Loan Exhibition of Historical Portraits and Relics, Metropolitan Opera House, New York City, April 17th to May 8th, 1889',
-    authors:
-     [ 'New York (N.Y.). Committee on Centennial Celebration of the Inauguration of Washington, 1789' ],
-    publishedDate: '1889',
-    industryIdentifiers: [ { type: 'OTHER', identifier: 'HARVARD:FL2UTE' } ],
-    readingModes: { text: false, image: true },
-    pageCount: 178,
-    printType: 'BOOK',
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: '0.1.1.0.full.1',
-    panelizationSummary: { containsEpubBubbles: false, containsImageBubbles: false },
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=lpMMAAAAYAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=lpMMAAAAYAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=lpMMAAAAYAAJ&printsec=frontcover&dq=inauthor:new+york&hl=&cd=9&source=gbs_api',
-    infoLink:
-     'https://play.google.com/store/books/details?id=lpMMAAAAYAAJ&source=gbs_api',
-    canonicalVolumeLink: 'https://market.android.com/details?id=book-lpMMAAAAYAAJ' } }
-{ info:
-  { title:
-     'Annual Report of the Attorney General of the State of New York',
-    authors: [ "New York (State). Attorney General's Office" ],
-    publishedDate: '1964',
-    description:
-     'Includes a section called Opinions of the Attorney General.',
-    industryIdentifiers: [ { type: 'OTHER', identifier: 'HARVARD:HL5KJV' } ],
-    readingModes: { text: false, image: true },
-    printType: 'BOOK',
-    categories: [ "Attorneys general's opinions" ],
-    maturityRating: 'NOT_MATURE',
-    allowAnonLogging: false,
-    contentVersion: '0.0.2.0.full.1',
-    panelizationSummary: { containsEpubBubbles: false, containsImageBubbles: false },
-    imageLinks:
-     { smallThumbnail:
-        'http://books.google.com/books/content?id=jkQQAAAAYAAJ&printsec=frontcover&img=1&zoom=5&edge=curl&source=gbs_api',
-       thumbnail:
-        'http://books.google.com/books/content?id=jkQQAAAAYAAJ&printsec=frontcover&img=1&zoom=1&edge=curl&source=gbs_api' },
-    language: 'en',
-    previewLink:
-     'http://books.google.com/books?id=jkQQAAAAYAAJ&pg=PA177&dq=inauthor:new+york&hl=&cd=10&source=gbs_api',
-    infoLink:
-     'https://play.google.com/store/books/details?id=jkQQAAAAYAAJ&source=gbs_api',
-    canonicalVolumeLink: 'https://market.android.com/details?id=book-jkQQAAAAYAAJ' } }
